+++ dnSpy/dnSpy.Contracts.Debugger/Engine/DbgEngineMessage.cs
@@ -61,14 +61,14 @@ namespace dnSpy.Contracts.Debugger.Engine {
 		/// <summary>
 		/// Gets the process id
 		/// </summary>
-		public ulong ProcessId { get; }
+		public int ProcessId { get; }
 
 		/// <summary>
 		/// Constructor
 		/// </summary>
 		/// <param name="processId">Process id</param>
 		/// <param name="messageFlags">Message flags</param>
-		public DbgMessageConnected(ulong processId, DbgEngineMessageFlags messageFlags) : base(messageFlags) => ProcessId = processId;
+		public DbgMessageConnected(int processId, DbgEngineMessageFlags messageFlags) : base(messageFlags) => ProcessId = processId;
 
 		/// <summary>
 		/// Constructor
